Methods

Index

StableDQMC.calc_Bchain

StableDQMC.calc_Bchain_qr

StableDQMC.calc_Bchain_svd

StableDQMC.calc_tdgf_qr

StableDQMC.calc_tdgf_svd

StableDQMC.fact_mult

StableDQMC.fact_mult

StableDQMC.gesdd

StableDQMC.gesdd!

StableDQMC.gesvd

StableDQMC.gesvd!

StableDQMC.inv!

StableDQMC.inv_one_plus

StableDQMC.inv_one_plus

StableDQMC.inv_one_plus

StableDQMC.inv_one_plus!

StableDQMC.inv_one_plus!

StableDQMC.inv_one_plus!

StableDQMC.inv_one_plus_loh

StableDQMC.inv_one_plus_loh

StableDQMC.inv_one_plus_loh!

StableDQMC.inv_one_plus_loh!

StableDQMC.inv_sum

StableDQMC.inv_sum

StableDQMC.inv_sum!

StableDQMC.inv_sum!

StableDQMC.inv_sum_loh

StableDQMC.inv_sum_loh

StableDQMC.inv_sum_loh!

StableDQMC.inv_sum_loh!

StableDQMC.svd_inv_one_plus

StableDQMC.svd_inv_one_plus_loh

StableDQMC.svd_inv_sum

StableDQMC.svd_inv_sum_loh

StableDQMC.udt

StableDQMC.udt!

StableDQMC.udt_inv_one_plus

StableDQMC.udt_inv_one_plus

StableDQMC.udt_inv_one_plus_loh

StableDQMC.udt_inv_sum

StableDQMC.udt_inv_sum_loh

Documentation

StableDQMC.calc_Bchain — Method

calc_Bchain(B,N) -> (R, svs)

Calculate B^N as BBBB...*B from right to left.

Returns the result R and the singular values (column of svs) of all intermediate powers (rows of svs).

StableDQMC.calc_Bchain_qr — Method

Same as calc_Bchain but stabilizes the matrix products by intermediate matrix decompositions.

StableDQMC.calc_Bchain_svd — Method

Same as calc_Bchain but stabilizes the matrix products by intermediate matrix decompositions.

StableDQMC.calc_tdgf_qr — Method

Calculate fake "G(τ, 0)", i.e. [B^-N1 + B^N2]^-1

StableDQMC.calc_tdgf_svd — Method

Calculate fake "G(τ, 0)", i.e. [B^-N1 + B^N2]^-1

StableDQMC.fact_mult — Method

fact_mult(A::SVD, B::SVD) -> SVD

Stabilized multiplication of two SVD decompositions. Returns a SVD factorization object.

StableDQMC.fact_mult — Method

fact_mult(A::UDT, B::UDT) -> UDT

Stabilized multiplication of two UDT decompositions. Returns a UDT factorization object.

StableDQMC.gesdd! — Method

Same as gesdd but saves space by overwriting the input matrix.

StableDQMC.gesdd — Method

Calculates the SVD deomposition of a matrix by employing a divide and conquer approach. Returns a SVD factorization object.

StableDQMC.gesvd! — Method

Same as gesvd but saves space by overwriting the input matrix.

StableDQMC.gesvd — Method

Calculates the SVD deomposition of a matrix. Returns a SVD factorization object.

StableDQMC.inv! — Method

inv!(res, F::UDT) -> res

Same as inv but writes result into preallocated res.

StableDQMC.inv_one_plus! — Method

inv_one_plus!(res, F::SVD) -> res

Stabilized calculation of [1 + USVt]^(-1):

Use one intermediate SVD decomposition.

Writes the result into res.

See svd_inv_one_plus for preallocation options.

StableDQMC.inv_one_plus! — Method

inv_one_plus!(res, A::UDT, Bdagger::UDT) -> res

Stabilized calculation of [1 + UlDlTl(UrDrTr)^†]^(-1). Writes the result into res.

See udt_inv_one_plus for preallocation options.

StableDQMC.inv_one_plus! — Method

inv_one_plus!(res, F::UDT) -> res

Same as inv_one_plus but stores the result in preallocated res.

StableDQMC.inv_one_plus — Method

inv_one_plus(F::SVD) -> AbstractMatrix

Stabilized calculation of [1 + USVt]^(-1):

Use one intermediate SVD decomposition.

See svd_inv_one_plus for preallocation options.

StableDQMC.inv_one_plus — Method

inv_one_plus(A::UDT, Bdagger::UDT) -> AbstractMatrix

Stabilized calculation of [1 + UlDlTl(UrDrTr)^†]^(-1).

See udt_inv_one_plus for preallocation options.

StableDQMC.inv_one_plus — Method

inv_one_plus(F::UDT) -> AbstractMatrix

Stabilized calculation of [1 + UDT]^(-1):

Use one intermediate UDT decomposition.

Faster but potentially less accurate than inv_one_plus_loh.

See udt_inv_one_plus for preallocation options.

StableDQMC.inv_one_plus_loh! — Method

inv_one_plus_loh!(res, F::SVD) -> res

Stabilized calculation of [1 + USVt]^(-1):

Separate scales larger and smaller than unity

Use two intermediate SVD decompositions.

Writes the result into res.

See svd_inv_one_plus_loh for preallocation options.

StableDQMC.inv_one_plus_loh! — Method

inv_one_plus_loh!(res, F::UDT) -> res

Stabilized calculation of [1 + UDT]^(-1):

Separate scales larger and smaller than unity

Use two intermediate UDT decompositions.

Writes the result into res.

See udt_inv_one_plus_loh for preallocation options.

StableDQMC.inv_one_plus_loh — Method

inv_one_plus_loh(F::SVD) -> AbstractMatrix

Stabilized calculation of [1 + USVt]^(-1):

Separate scales larger and smaller than unity

Use two intermediate SVD decompositions.

See svd_inv_one_plus_loh for preallocation options.

StableDQMC.inv_one_plus_loh — Method

inv_one_plus_loh(F::UDT) -> AbstractMatrix

Stabilized calculation of [1 + UDT]^(-1):

Separate scales larger and smaller than unity

Use two intermediate UDT decompositions.

See udt_inv_one_plus_loh for preallocation options.

StableDQMC.inv_sum! — Method

inv_sum!(res, A::SVD, B::SVD) -> res

Stabilized calculation of [UaSaVta + UbSbVtb]^(-1):

Use one intermediate SVD decompositions.

Writes the result into res.

See svd_inv_sum for preallocation options.

StableDQMC.inv_sum! — Method

inv_sum!(res, A::UDT, B::UDT) -> res

Stabilized calculation of [UaDaTa + UbDbTb]^(-1):

Use one intermediate UDT decompositions.

See udt_inv_sum for preallocation options.

StableDQMC.inv_sum — Method

inv_sum(A::SVD, B::SVD) -> res

Stabilized calculation of [UaSaVta + UbSbVtb]^(-1):

Use one intermediate SVD decompositions.

See svd_inv_sum for preallocation options.

StableDQMC.inv_sum — Method

inv_sum(A::UDT, B::UDT) -> AbstractMatrix

Stabilized calculation of [UaDaTa + UbDbTb]^(-1):

Use one intermediate UDT decompositions.

See udt_inv_sum for preallocation options.

StableDQMC.inv_sum_loh! — Method

inv_sum_loh!(res, A::SVD, B::SVD) -> res

Stabilized calculation of [UaSaVta + UbSbVtb]^(-1):

Separate scales larger and smaller than unity

Use two intermediate SVD decompositions.

Writes the result into res.

See svd_inv_sum_loh for preallocation options.

StableDQMC.inv_sum_loh! — Method

inv_sum_loh!(res, A::UDT, B::UDT) -> res

Stabilized calculation of [UaDaTa + UbDbTb]^(-1):

Separate scales larger and smaller than unity

Use two intermediate UDT decompositions.

Writes the result into res.

See udt_inv_sum_loh for preallocation options.

StableDQMC.inv_sum_loh — Method

inv_sum_loh(A::SVD, B::SVD) -> res

Stabilized calculation of [UaSaVta + UbSbVtb]^(-1):

Separate scales larger and smaller than unity

Use two intermediate SVD decompositions.

See svd_inv_sum_loh for preallocation options.

StableDQMC.inv_sum_loh — Method

inv_sum_loh(A::UDT, B::UDT) -> AbstractMatrix

Stabilized calculation of [UaDaTa + UbDbTb]^(-1):

Separate scales larger and smaller than unity

Use two intermediate UDT decompositions.

See udt_inv_sum_loh for preallocation options.

StableDQMC.svd_inv_one_plus — Method

svd_inv_one_plus(F::SVD) -> SVD

Stabilized calculation of [1 + USVt]^(-1):

Use one intermediate SVD decomposition.

Options for preallocation via keyword arguments:

t = similar(F.U)

a = similar(F.U)

c = similar(F.U)

StableDQMC.svd_inv_one_plus_loh — Method

svd_inv_one_plus_loh(F::SVD) -> SVD

Stabilized calculation of [1 + USVt]^(-1):

Separate scales larger and smaller than unity

Use two intermediate SVD decompositions.

Options for preallocation via keyword arguments:

l = similar(F.U)

r = similar(F.U)

Dp = similar(F.D)

Dm = similar(F.D)

StableDQMC.svd_inv_sum — Method

svd_inv_sum(A::SVD, B::SVD) -> SVD

Stabilized calculation of [UaSaVta + UbSbVtb]^(-1):

Use one intermediate SVD decompositions.

Options for preallocations via keyword arguments:

m1 = similar(A.U)

m2 = similar(A.U)

StableDQMC.svd_inv_sum_loh — Method

svd_inv_sum_loh(A::SVD, B::SVD) -> SVD

Stabilized calculation of [UaSaVta + UbSbVtb]^(-1):

Separate scales larger and smaller than unity

Use two intermediate SVD decompositions.

Options for preallocations via keyword arguments:

None (yet)

StableDQMC.udt! — Method

udv! is the same as svd, but saves space by overwriting the input A, instead of creating a copy.

StableDQMC.udt — Method

Compute the UDT decomposition of A and return an UDT object.

U, D, and T, can be obtained from the factorization F with F.U, F.D, and F.T such that A = U * Diagonal(D) * T.

Iterating the decomposition produces the components U, D, and V.

Note that T is upper triangular only up to permutations of columns of T.

StableDQMC.udt_inv_one_plus — Method

udt_inv_one_plus(A::UDT, Bdagger::UDT) -> UDT

Stabilized calculation of [1 + UlDlTl(UrDrTr)^†]^(-1). Returns and UDT factorization object.

Optional preallocations via keyword arguments:

tmp = similar(A.U)

tmp2 = similar(A.U)

tmp3 = similar(A.U)

StableDQMC.udt_inv_one_plus — Method

udt_inv_one_plus(F::UDT) -> UDT

Stabilized calculation of [1 + UDT]^(-1). Returns and UDT factorization object.

Optional preallocations via keyword arguments:

u = similar(F.U)

t = similar(F.T)

StableDQMC.udt_inv_one_plus_loh — Method

udt_inv_one_plus_loh(F::UDT) -> UDT

Stabilized calculation of [1 + UDT]^(-1):

Separate scales larger and smaller than unity

Use two intermediate UDT decompositions.

Options for preallocation via keyword arguments:

l = similar(F.U)

r = similar(F.U)

Dp = similar(F.D)

Dm = similar(F.D)

StableDQMC.udt_inv_sum — Method

udt_inv_sum(A::UDT, B::UDT) -> UDT

Stabilized calculation of [UaDaTa + UbDbTb]^(-1):

Use one intermediate UDT decompositions.

Optional preallocations via keyword arguments:

m2 = similar(A.U)

StableDQMC.udt_inv_sum_loh — Method

udt_inv_sum_loh(A::UDT, B::UDT) -> UDT

Stabilized calculation of [UaDaTa + UbDbTb]^(-1):

Separate scales larger and smaller than unity

Use two intermediate UDT decompositions.

Options for preallocations via keyword arguments:

mat2 = similar(A.U)

Dap = similar(A.D)

Dam = similar(A.D)

Dbp = similar(A.D)

Dbm = similar(A.D)